refactor the reverse pubsub in zedbox to be part of pubsub
