+++ pkg/pillar/pubsub/util.go
@@ -5,8 +5,11 @@ package pubsub
 
 import (
 	"encoding/json"
+	"fmt"
+	"net"
 	"reflect"
 	"strings"
+	"syscall"
 
 	"github.com/lf-edge/eve/pkg/pillar/base"
 )
@@ -60,4 +63,37 @@ func TypeToName(something interface{}) string {
 	t := reflect.TypeOf(something)
 	out := strings.Split(t.String(), ".")
 	return out[len(out)-1]
+}
+
+// ConnReadCheck waits till conn's fd is readable
+func ConnReadCheck(conn net.Conn) error {
+	var sysErr error
+
+	sysConn, ok := conn.(syscall.Conn)
+	if !ok {
+		return fmt.Errorf("Not syscall.Conn")
+	}
+	rawConn, err := sysConn.SyscallConn()
+	if err != nil {
+		return fmt.Errorf("Exception while getting rawConn: %s",
+			err)
+	}
+
+	err = rawConn.Read(func(fd uintptr) bool {
+		_, _, err := syscall.Recvfrom(int(fd), []byte{}, syscall.MSG_PEEK)
+		if err != nil {
+			if err == syscall.EAGAIN {
+				return false
+			}
+			//assign unknown error to syserr which will be handled later.
+			sysErr = fmt.Errorf("Unknown error from syscall.Recvfrom: %s",
+				err)
+		}
+		return true
+	})
+	if err != nil {
+		return fmt.Errorf("Exception from rawConn.Read: %s",
+			err)
+	}
+	return sysErr
 }
